Move code to new file multioutput.py
