+++ sklearn/tests/test_multiclass.py
@@ -13,7 +13,6 @@ from sklearn.utils.testing import assert_raise_message
 from sklearn.multiclass import OneVsRestClassifier
 from sklearn.multiclass import OneVsOneClassifier
 from sklearn.multiclass import OutputCodeClassifier
-from sklearn.multiclass import MultiOutputRegressor
 from sklearn.utils.multiclass import check_classification_targets, type_of_target
 from sklearn.utils import shuffle
 
@@ -26,7 +25,6 @@ from sklearn.linear_model import (LinearRegression, Lasso, ElasticNet, Ridge,
                                   Perceptron, LogisticRegression,
                                   SGDClassifier)
 from sklearn.tree import DecisionTreeClassifier, DecisionTreeRegressor
-from sklearn.ensemble import GradientBoostingRegressor
 from sklearn.model_selection import GridSearchCV
 from sklearn.pipeline import Pipeline
 from sklearn import svm
@@ -75,50 +73,6 @@ def test_ovr_fit_predict():
     ovr = OneVsRestClassifier(MultinomialNB())
     pred = ovr.fit(iris.data, iris.target).predict(iris.data)
     assert_greater(np.mean(iris.target == pred), 0.65)
-
-
-def test_multi_target_regression():
-    X, y = datasets.make_regression(n_targets=3)
-    X_train, y_train = X[:50], y[:50]
-    X_test, y_test = X[50:], y[50:]
-
-    references = np.zeros_like(y_test)
-    for n in range(3):
-        rgr = GradientBoostingRegressor(random_state=0)
-        rgr.fit(X_train, y_train[:, n])
-        references[:,n] = rgr.predict(X_test)
-
-    rgr = MultiOutputRegressor(GradientBoostingRegressor(random_state=0))
-    rgr.fit(X_train, y_train)
-    y_pred = rgr.predict(X_test)
-
-    assert_almost_equal(references, y_pred)
-
-
-def test_multi_target_regression_one_target():
-    # Test multi target regression raises
-    X, y = datasets.make_regression(n_targets=1)
-    X_train, y_train = X[:50], y[:50]
-    X_test, y_test = X[50:], y[50:]
-
-    rgr = MultiOutputRegressor(GradientBoostingRegressor(random_state=0))
-    assert_raises(ValueError, rgr.fit, X_train, y_train)
-
-
-def test_multi_target_sparse_regression():
-    X, y = datasets.make_regression(n_targets=3)
-    X_train, y_train = X[:50], y[:50]
-    X_test, y_test = X[50:], y[50:]
-
-    for sparse in [sp.csr_matrix, sp.csc_matrix, sp.coo_matrix, sp.dok_matrix,
-                   sp.lil_matrix]:
-        rgr = MultiOutputRegressor(Lasso(random_state=0))
-        rgr_sparse = MultiOutputRegressor(Lasso(random_state=0))
-
-        rgr.fit(X_train, y_train)
-        rgr_sparse.fit(sparse(X_train), y_train)
-
-        assert_almost_equal(rgr.predict(X_test), rgr_sparse.predict(sparse(X_test)))
 
 
 def test_ovr_partial_fit():
